feat(api): move `apiRequest` to lib/server for server-side usage and update it to forward cookies
